fix rename file (case)
