+++ src/components/_activities.js
